veritas/veritas_full_repo/backend/requirements.txt
fastapi==0.116.1
slowapi==0.1.9
limits>=4.2,<5
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator>=2.0.0
psycopg[binary]==3.2.10
python-multipart==0.0.20
redis==6.4.0
rq==2.6.0

paramiko==3.4.0

pytest==8.3.5
httpx==0.28.1
prometheus-client==0.22.1

PyYAML>=6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4